preen & clean
remove executable bits from *.yaml files
s/example.com/sad-emerald.state-space-network/
configtx.yaml, prepare as if the kafka stanza would be operable with reasonable sad-emerald-themed hostnames

+++ examples/sad-emerald/1.byfn.sh
@@ -77,14 +77,20 @@ cryptogen generate --config=./crypto-config.yaml --output ./crypto-config
     for kind in ca tlsca; do
       for dir in $(find $directory -type d -name $kind); do (
           cd $dir && for file in $kind.*.pem ; do
-            ln -vfs *_sk ${file%.pem}.key
+              # $file might be the unexpandable '*.pem'
+              if test -e "$file" ; then
+                  ln -vfs *_sk ${file%.pem}.key
+              fi
           done 
       ); done
     done
     for dir in $(find $directory -type d -name admincerts); do (
         if [[ -d ${dir%/*}/keystore ]] ; then
             cd $dir && for file in *.pem ; do
-                ( cd ../keystore && ln -vfs *_sk ${file%.pem}.key )
+                # $file might be the unexpandable '*.pem'
+                if test -e "$file" ; then
+                    ( cd ../keystore && ln -vfs *_sk ${file%.pem}.key )
+                    fi
             done 
          fi
     ); done
